Python/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Semantic Kernel (latest version compatible with Python 3.13)
semantic-kernel>=1.14.0

# Azure OpenAI
openai>=1.12.0

# Azure Identity for M365 Copilot authentication (future use)
azure-identity>=1.15.0

# HTTP client
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.3

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0

# M365 Agents SDK packages (now integrated)
microsoft-agents-hosting-fastapi
microsoft-agents-hosting-core
microsoft-agents-authentication-msal
microsoft-agents-activity

# M365 Copilot SDK (for future M365 Copilot integration)
microsoft-agents-m365copilot-beta
#microsoft-agents-m365copilot-core
